Analyzer.AnalyzedType

Description

Types representation during the analysis process.

Documentation

data Type

Type representation during the analysis process.

Constructors

TInt
32-bit/64-bit (depending on the machine) integer type.

TBool
Boolean type.

TString
String type.

TArray Type Int
Array type, it contains the length of the array and its elements type.

TFunction FunctionType
Function type, see FunctionType.

TNil
Null type.

data FunctionType

Function type, it contains the result of the function (which can be void if the result is equal to Nothing) and its parameters types.

Constructors

FunctionType

Fields

parameters :: [Type]

returnType :: MaybeVoid Type