Interpreter.Result

Interpretation result

type Result s a = ExceptT Err (StateT (Env s) (ST s)) a

Represents the result of interpretation.

lift2 :: ST s a -> Result s a

Result value

type ResultValue = Either Err

data Err

Represents unsuccessful interpretation.

Constructors

DivisionByZero
Division by zero error.

IndexOutOfRange Int Int
Index out of bounds error, contains info about index and array length.

NoReturn
No return error.

Npe
Null dereference error (happens when trying to call the nil as a regular function).

Panic Text
Panic error (happens when calling panic stdlib function).

UnexpectedError
Unexpected error, this type of errors must never happen.

State

Environment

data Env s

Constructors

Env

Fields

_funcs :: Map Identifier (STRef s Function, Scope s)

_funcScopes :: [FuncScope s]

_accumulatedOutput :: AccOut

type AccOut = [Text]

Accumulated output (every element is a text printed to the stdout).

emptyEnv :: Env s

Function scope

newtype FuncScope s

Constructors

FuncScope

Fields

_scopes :: [Scope s]

Scope

newtype Scope s

Scope contains identifiers mapped to their types.

Constructors

Scope

Fields

_vars :: Map Identifier (STRef s (RuntimeValue s))

emptyScope :: Scope s

Create empty Scope.

Runtime value

data RuntimeValue s

Represents runtime value of the calculated expression.

Constructors

ValInt Int
Int value.

ValBool Bool
Boolean value.

ValString Text
String value.

ValArray [RuntimeValue s]
Array value.

ValFunction FunctionValue (Scope s)
Function value.

Optics

funcs :: forall s. Lens' (Env s) (Map Identifier (STRef s Function, Scope s))

funcScopes :: forall s. Lens' (Env s) [FuncScope s]

accumulatedOutput :: forall s. Lens' (Env s) AccOut

scopes :: forall s s. Iso (FuncScope s) (FuncScope s) [Scope s] [Scope s]

vars :: forall s s. Iso (Scope s) (Scope s) (Map Identifier (STRef s (RuntimeValue s))) (Map Identifier (STRef s (RuntimeValue s)))

var :: Applicative f => Int -> Int -> Identifier -> LensLike' f (Env s) (Maybe (STRef s (RuntimeValue s)))

Evaluated state

Environment

data Env'

Constructors

Env'

Fields

_funcs' :: Map Identifier Function

_funcScopes' :: [FuncScope']

_accumulatedOutput' :: AccOut

evalEnv :: Env s -> ST s Env'

Function scope

newtype FuncScope'

Constructors

FuncScope'

Fields

_scopes' :: [Scope']

evalFuncScope :: FuncScope s -> ST s FuncScope'

Scope

newtype Scope'

Constructors

Scope'

Fields

_vars' :: Map Identifier RuntimeValue'

evalScope :: Scope s -> ST s Scope'

Runtime value

data RuntimeValue'

Represents runtime value of the calculated expression.

Constructors

ValInt' Int
Int value.

ValBool' Bool
Boolean value.

ValString' Text
String value.

ValArray' [RuntimeValue']
Array value.

ValFunction' FunctionValue
Function value.

evalRuntimeValue :: RuntimeValue s -> RuntimeValue'

unevalRuntimeValue :: RuntimeValue' -> RuntimeValue s

Optics

funcs' :: Lens' Env' (Map Identifier Function)

funcScopes' :: Lens' Env' [FuncScope']

accumulatedOutput' :: Lens' Env' AccOut

scopes' :: Iso' FuncScope' [Scope']

vars' :: Iso' Scope' (Map Identifier RuntimeValue')