Parser.Parser

Description

Provides parser that produces AST.

Parser

parse :: Text -> Maybe Program

Parser entry point.

Program (top-level) parsers

programP :: Parser Program

Program parser.

topLevelDeclP :: Parser (Either VarDecl FunctionDef)

Top-level declaration parser, it parses either var declaration or function definition.

functionDefP :: Parser FunctionDef

Function definition parser.

Expressions parsers

expressionP :: Parser Expression

Expression parser.

termExpressionP :: Parser Expression

Terminal expression parser, it's terminal in terms of makeExprParser parser.

opsTable :: [[Operator Parser Expression]]

Operators table, contains all operator parsers and their fixity.

Operators

binaryOp :: Text -> BinaryOp -> Operator Parser Expression

Utility function, that takes operator symbol, binary operator constructor and gives new binary operator in return.

unaryOp :: Text -> UnaryOp -> Operator Parser Expression

Utility function, that takes operator symbol, unary operator constructor and gives new unary operator in return.

funcCallOp :: Parser (Expression -> Expression)

Function call operator.

arrayAccessByIndexOp :: Parser (Expression -> Expression)

Array access by index operator.

Types parsers

typeP :: Parser Type

Type parser.

arrayTypeP :: Parser ArrayType

Array type parser.

functionTypeP :: Parser FunctionType

Function type parser.

Statements parsers

statementP :: Parser Statement

Statement parser.

stmtReturnP :: Parser Statement

Return statement parser.

stmtForGoToP :: Parser Statement

For goto statement parser.

stmtForP :: Parser Statement

For statement parser.

varDeclP :: Parser VarDecl

Var declaration parser.

ifElseP :: Parser IfElse

If-else parser.

blockP :: Parser Block

Block parser.

simpleStmtP :: Parser SimpleStmt

Simple statement parser.

stmtAssignmentP :: Parser SimpleStmt

Assignment statement parser.

stmtIncDecP :: Parser SimpleStmt

Increment or decrement statement parser.

stmtShortVarDeclP :: Parser SimpleStmt

Short var declaration statement parser.

stmtExpressionP :: Parser SimpleStmt

Expression statement parser.

lvalueP :: Parser Lvalue

Lvalue parser.

Values parsers

valueP :: Parser Value

Value parser.

arrayValP :: Parser ArrayValue

Array value parser.

functionValP :: Parser FunctionValue

Function value parser.

functionP :: Parser Function

Nameless function parser.

Utils

choice' :: (Foldable f, MonadParsec e s m, Functor f) => f (m a) -> m a

Choice between elements parser with built-in backtracking support.

eitherP' :: MonadParsec e s m => m a -> m b -> m (Either a b)

Combine two alternatives with built-in backtracking support.

optional' :: MonadParsec e s m => m a -> m (Maybe a)

Optional element parser with built-in backtracking support.

maybeVoid :: MonadParsec e s m => m a -> m (MaybeVoid a)